+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add message type gas statistics #260
 - Add calls to getAssetMetadata where explicitly needed #283
 - Add buttons for governance voting #242
+- Add fee stats to transactions chart on dashboard #301
 
 ## 2.6.0